make Ordered support skimming (eg. TOP) limits
Skimming limits like TOP now require a unique key in the order_by clause
for stability, but Ordered only uses the position column in its order_by
clauses, while Result classes using Ordered generally have a compound
unique constraint on the grouping column and position column.

Append the grouping column to the end of the Ordered order_by clauses so
that it uses the full unique constraint and satisfy the skimming limit
stability requirements. This is harmless as the rows will still be
ordered on the position column.